Ouroboros.Consensus.ByronSpec.Ledger

Documentation

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family Validated x

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

data family LedgerState blk

Ledger state associated with a block

data family BlockConfig blk

Static configuration required to work with this type of blocks

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

data family Header blk

data family Ticked st

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

data ByronSpecGenesis

The equivalent of the genesis config for the abstract ledger

Constructors

ByronSpecGenesis

Fields

byronSpecGenesisDelegators ∷ Set VKeyGenesis

byronSpecGenesisInitUtxo ∷ UTxO

byronSpecGenesisInitPParams ∷ PParams

byronSpecGenesisSecurityParam ∷ BlockCount

byronSpecGenesisSlotLength ∷ Natural

Slot length

The Byron spec itself does not talk about slot length at all. Here we record it primarily to support the relation between the spec and the real implementation. For this reason we choose the same representation as the real PBFT does (ppSlotDuration in ProtocolParameters).

type ByronSpecHeader = Header ByronSpecBlock

data ByronSpecBlock

Block according to the Byron spec

Just like we do for ByronBlock, we cache the header hash. In addition, we also add the BlockNo, as this is entirely absent from the spec but we need it for the HasHeader abstraction, which is ubiquitous in ouroboros-consensus and -network.

Constructors

ByronSpecBlock

Fields

byronSpecBlock ∷ Block

byronSpecBlockNo ∷ BlockNo

byronSpecBlockHash ∷ Hash

newtype ByronSpecLedgerError

Constructors

ByronSpecLedgerError

Fields

unByronSpecLedgerError ∷ [PredicateFailure CHAIN]

initByronSpecLedgerState ∷ ByronSpecGenesis → LedgerState ByronSpecBlock

newtype ByronSpecGenTxErr

Transaction errors

We don't distinguish these from any other kind of CHAIN failure.

Constructors

ByronSpecGenTxErr

Fields

unByronSpecGenTxErr ∷ [PredicateFailure CHAIN]

data ByronSpecGenTx

Generalized transaction

The spec doesn't have a type for this, instead splitting the block body into separate lists

Constructors

ByronSpecGenTxDCert DCert

ByronSpecGenTxTx Tx

ByronSpecGenTxUProp UProp

ByronSpecGenTxVote Vote

forgeByronSpecBlock ∷ BlockNo → SlotNo → Ticked (LedgerState ByronSpecBlock) → [Validated (GenTx ByronSpecBlock)] → VKey → ByronSpecBlock