Ouroboros.Consensus.ByronSpec.Ledger.Ledger

Documentation

newtype ByronSpecLedgerError

Constructors

ByronSpecLedgerError

Fields

unByronSpecLedgerError ∷ [[PredicateFailure CHAIN]]

initByronSpecLedgerState ∷ ByronSpecGenesis → LedgerState ByronSpecBlock

Type family instances

data family LedgerState blk ∷ Type

Ledger state associated with a block

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

Orphan instances

CommonProtocolParams ByronSpecBlock

Instance details

Methods

maxHeaderSize ∷ LedgerState ByronSpecBlock → Word32

maxTxSize ∷ LedgerState ByronSpecBlock → Word32

UpdateLedger ByronSpecBlock

Instance details

Eq (LedgerState ByronSpecBlock)

Instance details

Methods

(==) ∷ LedgerState ByronSpecBlock → LedgerState ByronSpecBlock → Bool

(/=) ∷ LedgerState ByronSpecBlock → LedgerState ByronSpecBlock → Bool

Eq (Ticked (LedgerState ByronSpecBlock))

Instance details

Methods

(==) ∷ Ticked (LedgerState ByronSpecBlock) → Ticked (LedgerState ByronSpecBlock) → Bool

(/=) ∷ Ticked (LedgerState ByronSpecBlock) → Ticked (LedgerState ByronSpecBlock) → Bool

Show (LedgerState ByronSpecBlock)

Instance details

Methods

showsPrec ∷ Int → LedgerState ByronSpecBlock → ShowS

show ∷ LedgerState ByronSpecBlock → String

showList ∷ [LedgerState ByronSpecBlock] → ShowS

Show (Ticked (LedgerState ByronSpecBlock))

Instance details

Methods

showsPrec ∷ Int → Ticked (LedgerState ByronSpecBlock) → ShowS

show ∷ Ticked (LedgerState ByronSpecBlock) → String

showList ∷ [Ticked (LedgerState ByronSpecBlock)] → ShowS

Generic (LedgerState ByronSpecBlock)

Instance details

Associated Types

type Rep (LedgerState ByronSpecBlock) ∷ Type → Type

Methods

from ∷ LedgerState ByronSpecBlock → Rep (LedgerState ByronSpecBlock) x

to ∷ Rep (LedgerState ByronSpecBlock) x → LedgerState ByronSpecBlock

NoUnexpectedThunks (LedgerState ByronSpecBlock)

Instance details

Methods

noUnexpectedThunks ∷ [String] → LedgerState ByronSpecBlock → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → LedgerState ByronSpecBlock → IO ThunkInfo

showTypeOf ∷ Proxy (LedgerState ByronSpecBlock) → String

NoUnexpectedThunks (Ticked (LedgerState ByronSpecBlock))

Instance details

Methods

noUnexpectedThunks ∷ [String] → Ticked (LedgerState ByronSpecBlock) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → Ticked (LedgerState ByronSpecBlock) → IO ThunkInfo

showTypeOf ∷ Proxy (Ticked (LedgerState ByronSpecBlock)) → String

Serialise (LedgerState ByronSpecBlock)

Instance details

Methods

encode ∷ LedgerState ByronSpecBlock → Encoding

decode ∷ Decoder s (LedgerState ByronSpecBlock)

encodeList ∷ [LedgerState ByronSpecBlock] → Encoding

decodeList ∷ Decoder s [LedgerState ByronSpecBlock]

GetTip (LedgerState ByronSpecBlock)

Instance details

Methods

getTip ∷ LedgerState ByronSpecBlock → Point (LedgerState ByronSpecBlock)

GetTip (Ticked (LedgerState ByronSpecBlock))

Instance details

Methods

getTip ∷ Ticked (LedgerState ByronSpecBlock) → Point (Ticked (LedgerState ByronSpecBlock))

IsLedger (LedgerState ByronSpecBlock)

Instance details

Associated Types

type LedgerErr (LedgerState ByronSpecBlock) ∷ Type

Methods

applyChainTick ∷ LedgerCfg (LedgerState ByronSpecBlock) → SlotNo → LedgerState ByronSpecBlock → Ticked (LedgerState ByronSpecBlock)

ApplyBlock (LedgerState ByronSpecBlock) ByronSpecBlock

Instance details

Methods

applyLedgerBlock ∷ LedgerCfg (LedgerState ByronSpecBlock) → ByronSpecBlock → Ticked (LedgerState ByronSpecBlock) → Except (LedgerErr (LedgerState ByronSpecBlock)) (LedgerState ByronSpecBlock)

reapplyLedgerBlock ∷ LedgerCfg (LedgerState ByronSpecBlock) → ByronSpecBlock → Ticked (LedgerState ByronSpecBlock) → LedgerState ByronSpecBlock