Ouroboros.Consensus.ByronSpec.Ledger.Mempool

Type family instances

data family GenTx blk

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family Validated x

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

Orphan instances

LedgerSupportsMempool ByronSpecBlock

Instance details

Methods

txInvariant ∷ GenTx ByronSpecBlock → Bool

applyTx ∷ LedgerConfig ByronSpecBlock → WhetherToIntervene → SlotNo → GenTx ByronSpecBlock → TickedLedgerState ByronSpecBlock → Except (ApplyTxErr ByronSpecBlock) (TickedLedgerState ByronSpecBlock, Validated (GenTx ByronSpecBlock))

reapplyTx ∷ LedgerConfig ByronSpecBlock → SlotNo → Validated (GenTx ByronSpecBlock) → TickedLedgerState ByronSpecBlock → Except (ApplyTxErr ByronSpecBlock) (TickedLedgerState ByronSpecBlock)

txsMaxBytes ∷ TickedLedgerState ByronSpecBlock → Word32

txInBlockSize ∷ GenTx ByronSpecBlock → Word32

txForgetValidated ∷ Validated (GenTx ByronSpecBlock) → GenTx ByronSpecBlock

Show (GenTx ByronSpecBlock)

Instance details

Methods

showsPrec ∷ Int → GenTx ByronSpecBlock → ShowS

show ∷ GenTx ByronSpecBlock → String

showList ∷ [GenTx ByronSpecBlock] → ShowS

Show (Validated (GenTx ByronSpecBlock))

Instance details

Methods

showsPrec ∷ Int → Validated (GenTx ByronSpecBlock) → ShowS

show ∷ Validated (GenTx ByronSpecBlock) → String

showList ∷ [Validated (GenTx ByronSpecBlock)] → ShowS

Generic (GenTx ByronSpecBlock)

Instance details

Associated Types

type Rep (GenTx ByronSpecBlock) ∷ Type → Type

Methods

from ∷ GenTx ByronSpecBlock → Rep (GenTx ByronSpecBlock) x

to ∷ Rep (GenTx ByronSpecBlock) x → GenTx ByronSpecBlock

Generic (Validated (GenTx ByronSpecBlock))

Instance details

Associated Types

type Rep (Validated (GenTx ByronSpecBlock)) ∷ Type → Type

Methods

from ∷ Validated (GenTx ByronSpecBlock) → Rep (Validated (GenTx ByronSpecBlock)) x

to ∷ Rep (Validated (GenTx ByronSpecBlock)) x → Validated (GenTx ByronSpecBlock)

Serialise (GenTx ByronSpecBlock)

Instance details

Methods

encode ∷ GenTx ByronSpecBlock → Encoding

decode ∷ Decoder s (GenTx ByronSpecBlock)

encodeList ∷ [GenTx ByronSpecBlock] → Encoding

decodeList ∷ Decoder s [GenTx ByronSpecBlock]

NoThunks (GenTx ByronSpecBlock)

Instance details

Methods

noThunks ∷ Context → GenTx ByronSpecBlock → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → GenTx ByronSpecBlock → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (GenTx ByronSpecBlock) → String

NoThunks (Validated (GenTx ByronSpecBlock))

Instance details

Methods

noThunks ∷ Context → Validated (GenTx ByronSpecBlock) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Validated (GenTx ByronSpecBlock) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Validated (GenTx ByronSpecBlock)) → String