Ouroboros.Consensus.ByronSpec.Ledger.Rules

Description

Thin wrapper around the Byron spec rules

Intended for qualified import

import qualified Ouroboros.Consensus.ByronSpec.Ledger.Rules as Rules

Ledger

applyChainTick ∷ ByronSpecGenesis → Slot → State CHAIN → State CHAIN

Chain tick

There isn't something in the spec that directly corresponds to this, so we have to combine a few different things:

Apply the epoch rule (update the update state)

Apply any scheduled delegations

Set the slot number

This matches quite closely what applyChainTick in Ouroboros.Consensus.Ledger.Byron.Auxiliary does.

Lift STS transition rules to the chain level

liftCHAIN ∷ ByronSpecGenesis → LiftedRule CHAIN

Apply a block

This is a "trivial" (identity) lift.

liftSDELEG ∷ ByronSpecGenesis → LiftedRule SDELEG

Apply delegation certificate

liftUTXOW ∷ ByronSpecGenesis → LiftedRule UTXOW

Apply transaction

liftUPIREG ∷ ByronSpecGenesis → LiftedRule UPIREG

Apply update proposal

liftUPIVOTE ∷ ByronSpecGenesis → LiftedRule UPIVOTE

Apply update vote

STS initial rules

initStateCHAIN ∷ ByronSpecGenesis → State CHAIN

Rule context (exported for the benefit of the tests

data RuleContext sts

Context required to apply a rule to the top-level CHAIN state

The environment for these rules pull in some information from the (abstract) genesis config and some information from the state; the reason is that although some values come from the state, as far as these rules are concerned, they are constants.

Constructors

RuleContext

Fields

getRuleState ∷ GetChainState (State sts)

modRuleState ∷ ModChainState (State sts)

liftFailure ∷ PredicateFailure sts → PredicateFailure CHAIN

getRuleEnv ∷ State CHAIN → Environment sts

ctxtCHAIN ∷ ByronSpecGenesis → RuleContext CHAIN

ctxtEPOCH ∷ ByronSpecGenesis → RuleContext EPOCH

ctxtDELEG ∷ ByronSpecGenesis → RuleContext DELEG

ctxtSDELEG ∷ ByronSpecGenesis → RuleContext SDELEG

ctxtUTXOW ∷ ByronSpecGenesis → RuleContext UTXOW

ctxtUPIREG ∷ ByronSpecGenesis → RuleContext UPIREG

ctxtUPIVOTE ∷ ByronSpecGenesis → RuleContext UPIVOTE