Ouroboros.Consensus.ByronSpec.Ledger

Documentation

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family LedgerState blk ∷ Type

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

data family Header blk ∷ Type

data ByronSpecGenesis

The equivalent of the genesis config for the abstract ledger

Constructors

ByronSpecGenesis

Fields

byronSpecGenesisDelegators ∷ Set VKeyGenesis

byronSpecGenesisInitUtxo ∷ UTxO

byronSpecGenesisInitPParams ∷ PParams

byronSpecGenesisSecurityParam ∷ BlockCount

byronSpecGenesisSlotLength ∷ Natural

Slot length

The Byron spec itself does not talk about slot length at all. Here we record it primarily to support the relation between the spec and the real implementation. For this reason we choose the same representation as the real PBFT does (ppSlotDuration in ProtocolParameters).

type ByronSpecHeader = Header ByronSpecBlock

data ByronSpecBlock

Block according to the Byron spec

Just like we do for ByronBlock, we cache the header hash. In addition, we also add the BlockNo, as this is entirely absent from the spec but we need it for the HasHeader abstraction, which is ubiquitous in ouroboros-consensus and -network.

Constructors

ByronSpecBlock

Fields

byronSpecBlock ∷ Block

byronSpecBlockNo ∷ BlockNo

byronSpecBlockHash ∷ Hash

newtype ByronSpecLedgerError

Constructors

ByronSpecLedgerError

Fields

unByronSpecLedgerError ∷ [[PredicateFailure CHAIN]]

initByronSpecLedgerState ∷ ByronSpecGenesis → LedgerState ByronSpecBlock

updateByronSpecLedgerStateKeepTip ∷ LedgerState ByronSpecBlock → State CHAIN → LedgerState ByronSpecBlock

updateByronSpecLedgerStateNewTip ∷ SlotNo → State CHAIN → LedgerState ByronSpecBlock

newtype ByronSpecGenTxErr

Transaction errors

We don't distinguish these from any other kind of CHAIN failure.

Constructors

ByronSpecGenTxErr

Fields

unByronSpecGenTxErr ∷ [[PredicateFailure CHAIN]]

data ByronSpecGenTx

Generalized transaction

The spec doesn't have a type for this, instead splitting the block body into separate lists

Constructors

ByronSpecGenTxDCert DCert

ByronSpecGenTxTx Tx

ByronSpecGenTxUProp UProp

ByronSpecGenTxVote Vote

forgeByronSpecBlock ∷ SlotNo → BlockNo → LedgerState ByronSpecBlock → [GenTx ByronSpecBlock] → VKey → ByronSpecBlock