Ouroboros.Consensus.ByronSpec.Ledger.Block

Documentation

data family BlockConfig blk

Static configuration required to work with this type of blocks

data ByronSpecBlock

Block according to the Byron spec

Just like we do for ByronBlock, we cache the header hash. In addition, we also add the BlockNo, as this is entirely absent from the spec but we need it for the HasHeader abstraction, which is ubiquitous in ouroboros-consensus and -network.

Constructors

ByronSpecBlock

Fields

byronSpecBlock ∷ Block

byronSpecBlockNo ∷ BlockNo

byronSpecBlockHash ∷ Hash

data family CodecConfig blk

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family Header blk

data family StorageConfig blk

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

type alias

type ByronSpecHeader = Header ByronSpecBlock