Ouroboros.Consensus.ByronSpec.Ledger.Block

Documentation

data ByronSpecBlock

Block according to the Byron spec

Just like we do for ByronBlock, we cache the header hash. In addition, we also add the BlockNo, as this is entirely absent from the spec but we need it for the HasHeader abstraction, which is ubiquitous in ouroboros-consensus and -network.

Constructors

ByronSpecBlock

Fields

byronSpecBlock ∷ Block

byronSpecBlockNo ∷ BlockNo

byronSpecBlockHash ∷ Hash

type ByronSpecHeader = Header ByronSpecBlock

data family Header blk ∷ Type

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks