Ouroboros.Consensus.ByronSpec.Ledger.GenTx

Description

Combine all Byron transaction types into single union type

Intended for qualified import

import           Ouroboros.Consensus.ByronSpec.Ledger.GenTx (ByronSpecGenTx(..), ByronSpecGenTxErr(..))
import qualified Ouroboros.Consensus.ByronSpec.Ledger.GenTx as GenTx

Documentation

data ByronSpecGenTx

Generalized transaction

The spec doesn't have a type for this, instead splitting the block body into separate lists

Constructors

ByronSpecGenTxDCert DCert

ByronSpecGenTxTx Tx

ByronSpecGenTxUProp UProp

ByronSpecGenTxVote Vote

newtype ByronSpecGenTxErr

Transaction errors

We don't distinguish these from any other kind of CHAIN failure.

Constructors

ByronSpecGenTxErr

Fields

unByronSpecGenTxErr ∷ [[PredicateFailure CHAIN]]

apply ∷ ByronSpecGenesis → ByronSpecGenTx → State CHAIN → Except ByronSpecGenTxErr (State CHAIN)

partition ∷ [ByronSpecGenTx] → ([DCert], [Tx], [UProp], [Vote])