Ouroboros.Consensus.ByronSpec.Ledger.Genesis

Description

Genesis config for the spec

Intended for qualified import

import           Ouroboros.Consensus.ByronSpec.Ledger.Genesis (ByronSpecGenesis)
import qualified Ouroboros.Consensus.ByronSpec.Ledger.Genesis as Genesis

Documentation

data ByronSpecGenesis

The equivalent of the genesis config for the abstract ledger

Constructors

ByronSpecGenesis

Fields

byronSpecGenesisDelegators ∷ Set VKeyGenesis

byronSpecGenesisInitUtxo ∷ UTxO

byronSpecGenesisInitPParams ∷ PParams

byronSpecGenesisSecurityParam ∷ BlockCount

modPBftThreshold ∷ (Double → Double) → ByronSpecGenesis → ByronSpecGenesis

modFeeParams ∷ ((Int, Int) → (Int, Int)) → ByronSpecGenesis → ByronSpecGenesis

Modify the a and b fee parameters

modUtxo ∷ (UTxO → UTxO) → ByronSpecGenesis → ByronSpecGenesis

modUtxoValues ∷ (Integer → Integer) → ByronSpecGenesis → ByronSpecGenesis

Adjust all values in the initial UTxO equally

modPParams ∷ (PParams → PParams) → ByronSpecGenesis → ByronSpecGenesis

Conversions

toChainEnv ∷ ByronSpecGenesis → Environment CHAIN

Derive CHAIN rule environment

fromChainEnv ∷ Environment CHAIN → ByronSpecGenesis

Construct genesis config from CHAIN environment

This doesn't make an awful lot of sense, but the abstract spec doesn't have a concept of a genesis config, and instead the CHAIN environment fulfills that role. In order to be able to reuse the test generators, we therefore also define a translation in the opposite direction.